beta

+++ telegram.py
@@ -3,8 +3,10 @@ import os
 
 from telethon import TelegramClient, events
 import re
-from youtube import youtube_download
+from youtube import youtube_download, get_yt_title
 from dotenv import load_dotenv
+from regex import is_youtube_link
+from group import search_messages
 
 load_dotenv()
 # Replace these values with your own API ID, API HASH, and bot token
@@ -20,11 +22,6 @@ client = TelegramClient('bot', api_id, api_hash).start(bot_token=bot_token)
 @client.on(events.NewMessage())
 async def main(event: events.newmessage.NewMessage.Event):
 
-    def is_youtube_link(text):
-        youtube_pattern = re.compile(r"(?:https?:\/\/)?(?:www\.)?(?:youtu\.be\/|youtube\.com\/watch\?v=)([\w-]+)")
-        match = youtube_pattern.match(text)
-        return bool(match)
-
     if event.is_private and event.text == '/start':
         if event.sender.last_name is not None:
             await event.respond(f'''Salam, **{event.sender.first_name}** **{event.sender.last_name}**!
@@ -37,7 +34,7 @@ Sadəcə istədiyin Youtube linkini göndər və arxana söykən. 😉
 
 ❕Əgər musiqi daha əvvəl başqası tərəfindən yüklənibsə, anında göndərəcəyəm.
 
-🗄 İndirilen müzikleri kanalımda (https://t.me/joinchat/AAAAAEpNIFaaE_CgT1vTYQ) depoluyorum.''')
+🗄 Yüklənən musiqiləri [kanalımda](https://t.me/musicalaze) depolayıram.''')
         else:
             await event.respond(f'''Salam, **{event.sender.first_name}** !
 Mən göndərəcəyin Youtube linkini sənə MP3 formatında qaytarmaq üçün assistantam.
@@ -49,25 +46,23 @@ Sadəcə istədiyin Youtube linkini göndər və arxana söykən. 😉
 
 ❕Əgər musiqi daha əvvəl başqası tərəfindən yüklənibsə, anında göndərəcəyəm.
 
-🗄 İndirilen müzikleri kanalımda (https://t.me/joinchat/AAAAAEpNIFaaE_CgT1vTYQ) depoluyorum.''')
+🗄 Yüklənən musiqiləri [kanalımda](https://t.me/musicalaze) depolayıram.''')
 
     else:
 
-        try:
-            if is_youtube_link(event.raw_text):
-                new_file, yt_title = await youtube_download(event)
-                await client.send_file(event.chat_id, new_file, caption=yt_title)
+        if is_youtube_link(event.raw_text):
+            yt_title = get_yt_title(event)
+            paired_message = await search_messages(yt_title)
+            if paired_message:
+                await client.forward_messages(event.chat_id, paired_message)
+            else:
+                new_file = await youtube_download(event)
+                sent_message = await client.send_file(event.chat_id, new_file, caption=yt_title)
+                await client.forward_messages(1859107525, sent_message)
 
                 os.remove(new_file)
-            else:
-                await client.send_message(event.chat_id, "Zəhmət olmasa yalnız youtube linki göndərin")
-
-        except Exception as e:
-            await client.send_message(event.chat_id, "Xəta baş verdi. Zəhmət olmasa yaş məhdudiyyəti olmayan Youtube linki göndərin.")
+        else:
+            await client.send_message(event.chat_id, "Zəhmət olmasa yalnız youtube linki göndərin")
 
 
-
-
-
-with client:
-    client.run_until_disconnected()
+client.run_until_disconnected()
